Moved getSpecialGeoIds to GeoLevel.
